+++ db/migrate/20260804000002_default_runs_to_no_temperature.rb
@@ -1,0 +1,9 @@
+class DefaultRunsToNoTemperature < ActiveRecord::Migration[8.0]
+  def up
+    change_column_default :completion_kit_runs, :temperature, nil
+  end
+
+  def down
+    change_column_default :completion_kit_runs, :temperature, 1.0
+  end
+end